+++ src/psm/Module/Install/Controller/InstallController.php
@@ -210,6 +210,7 @@ class InstallController extends AbstractController {
 			'level' => PSM_USER_ADMIN,
 			'pushover_key' => '',
 			'pushover_device' => '',
+			'carriersms' => '',
 		);
 
 		$validator = $this->container->get('util.user.validator');